Added Multiple Instance Functionality and Hex Path Passing
