+++ Attiny/ATtiny85/Source_Codes/DUTghdl/compile.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+gcc -c ghdlserver.c
+ghdl -a Utility_Package.vhdl && 
+ghdl -a Vhpi_Package.vhdl 
+
+
