+++ Attiny/ATtiny85/Source_Codes/DUTghdl/hex.txt
@@ -1,0 +1,5 @@
+:100000000EC013C012C011C010C00FC00EC00DC072
+:100010000CC00BC00AC009C008C007C006C011242C
+:100020001FBECFEDCDBF02D008C0EACF87B383603B
+:0E00300087BB18BA18BAC09AFDCFF894FFCF5C
+:00000001FF
