+++ src/createKicadLibrary.py
@@ -2,18 +2,25 @@ from Appconfig import Appconfig
 import re
 import os
 import xml.etree.cElementTree as ET
-from PyQt4 import QtGui
+from PyQt5 import QtWidgets
 
 
-class AutoSchematic(QtGui.QWidget):
+class AutoSchematic(QtWidgets.QWidget):
 
-    def __init__(self, modelname):
-        QtGui.QWidget.__init__(self)
+    def __init__(self, parent, modelname):
+        QtWidgets.QWidget.__init__(self)
+        self.parent = parent
         self.modelname = modelname.split('.')[0]
         self.template = Appconfig.kicad_lib_template.copy()
         self.xml_loc = Appconfig.xml_loc
         self.lib_loc = Appconfig.lib_loc
-        self.kicad_nghdl_lib = '/usr/share/kicad/library/eSim_Nghdl.lib'
+        if os.name == 'nt':
+            eSim_src = Appconfig.src_home
+            inst_dir = eSim_src.replace('\\eSim', '')
+            self.kicad_nghdl_lib = \
+                inst_dir + '/KiCad/share/kicad/library/eSim_Nghdl.lib'
+        else:
+            self.kicad_nghdl_lib = '/usr/share/kicad/library/eSim_Nghdl.lib'
         self.parser = Appconfig.parser_nghdl
 
     def createKicadLibrary(self):
@@ -26,16 +33,16 @@ class AutoSchematic(QtGui.QWidget):
             self.getPortInformation()
             self.createXML()
             self.createLib()
-        elif (xmlFound == self.xml_loc + '/Nghdl'):
+        elif (xmlFound == os.path.join(self.xml_loc, 'Nghdl')):
             print('Library already exists...')
-            ret = QtGui.QMessageBox.warning(
-                self, "Warning", '''<b>Library files for this model ''' +
-                '''already exist. Do you want to overwrite it?</b><br/>
+            ret = QtWidgets.QMessageBox.warning(
+                self.parent, "Warning", '''<b>Library files for this model''' +
+                ''' already exist. Do you want to overwrite it?</b><br/>
                 If yes press ok, else cancel it and ''' +
                 '''change the name of your vhdl file.''',
-                QtGui.QMessageBox.Ok, QtGui.QMessageBox.Cancel
+                QtWidgets.QMessageBox.Ok, QtWidgets.QMessageBox.Cancel
             )
-            if ret == QtGui.QMessageBox.Ok:
+            if ret == QtWidgets.QMessageBox.Ok:
                 print("Overwriting existing libraries")
                 self.getPortInformation()
                 self.createXML()
@@ -46,11 +53,11 @@ class AutoSchematic(QtGui.QWidget):
                 quit()
         else:
             print('Pre existing library...')
-            ret = QtGui.QMessageBox.critical(
-                self, "Error", '''<b>A standard library already exists ''' +
-                '''with this name.</b><br/><b>Please change the name ''' +
-                '''of your vhdl file and upload it again</b>''',
-                QtGui.QMessageBox.Ok
+            ret = QtWidgets.QMessageBox.critical(
+                self.parent, "Error", '''<b>A standard library already ''' +
+                '''exists with this name.</b><br/><b>Please change the ''' +
+                '''name of your vhdl file and upload it again</b>''',
+                QtWidgets.QMessageBox.Ok
             )
 
             # quit()
@@ -72,7 +79,7 @@ class AutoSchematic(QtGui.QWidget):
         print("changing directory to ", xmlDestination)
         os.chdir(xmlDestination)
 
-        root = ET.Element("model")
+        root = ET.Element("microcontroller")
         ET.SubElement(root, "name").text = self.modelname
         ET.SubElement(root, "type").text = "Nghdl"
         ET.SubElement(root, "node_number").text = str(len(self.portInfo))
@@ -88,6 +95,8 @@ class AutoSchematic(QtGui.QWidget):
                                     "Enter Input Load (default=1.0e-12)")
         ET.SubElement(param, "instance_id", default="1").text = (
                                     "Enter Instance ID (Between 0-99)")
+        ET.SubElement(param, "hex_path", default="/home/").text = (
+                                    "Path of your .hex file")
         tree = ET.ElementTree(root)
         tree.write(str(self.modelname) + '.xml')
         print("Leaving the directory ", xmlDestination)
@@ -224,23 +233,25 @@ class AutoSchematic(QtGui.QWidget):
 
         os.chdir(cwd)
         print('Leaving directory, ', self.lib_loc)
-        QtGui.QMessageBox.information(
-            self, "Library added",
+        QtWidgets.QMessageBox.information(
+            self.parent, "Library added",
             '''Library details for this model is added to the ''' +
             '''<b>eSim_Nghdl.lib</b> in the KiCad shared directory''',
-            QtGui.QMessageBox.Ok
+            QtWidgets.QMessageBox.Ok
         )
 
 
 class PortInfo:
     def __init__(self, model):
         self.modelname = model.modelname
-        self.model_loc = model.parser.get('NGSPICE', 'DIGITAL_MODEL')
+        self.model_loc = os.path.join(
+            model.parser.get('NGHDL', 'DIGITAL_MODEL'), 'ghdl'
+        )
         self.bit_list = []
         self.input_len = 0
 
     def getPortInfo(self):
-        info_loc = os.path.join(self.model_loc, self.modelname+'/DUTghdl/')
+        info_loc = os.path.join(self.model_loc, self.modelname + '/DUTghdl/')
         input_list = []
         output_list = []
         read_file = open(info_loc + 'connection_info.txt', 'r')
