+++ src/model_generation.py
@@ -1,7 +1,6 @@
-#!/usr/bin/python3
-
 import re
 import os
+from configparser import ConfigParser
 
 
 class ModelGeneration:
@@ -12,7 +11,19 @@ class ModelGeneration:
         print("Arguement is : ", file)
         self.fname = os.path.basename(file)
         print("VHDL filename is : ", self.fname)
-        self.home = os.path.expanduser("~")
+
+        if os.name == 'nt':
+            self.home = os.path.join('library', 'config')
+        else:
+            self.home = os.path.expanduser('~')
+
+        self.parser = ConfigParser()
+        self.parser.read(os.path.join(
+            self.home, os.path.join('.nghdl', 'config.ini')))
+        self.nghdl_home = self.parser.get('NGHDL', 'NGHDL_HOME')
+        self.release_dir = self.parser.get('NGHDL', 'RELEASE')
+        self.src_home = self.parser.get('SRC', 'SRC_HOME')
+        self.licensefile = self.parser.get('SRC', 'LICENSE')
 
         # #### Creating connection_info.txt file from vhdl file #### #
         read_vhdl = open(file, 'r')
@@ -35,7 +46,13 @@ class ModelGeneration:
                 item = re.sub("\(", " ", item, flags=re.I)      # noqa
                 item = re.sub("\)", " ", item, flags=re.I)      # noqa
                 item = re.sub(";", " ", item, flags=re.I)
-
+                if item.find(','):
+                    temp1 = item.split(",")
+                    item = "" + temp1[-1]
+                    temp2 = temp1[-1].split(":")
+                    for i in range(len(temp1) - 1):
+                        temp3 = temp1[i] + ":" + temp2[-1]
+                        scan_data.append(temp3.rstrip())
                 scan_data.append(item.rstrip())
                 scan_data = [_f for _f in scan_data if _f]
             elif start_flag == 0:
@@ -116,23 +133,23 @@ class ModelGeneration:
                 out_items = re.findall(
                     "OUT", line, re.MULTILINE | re.IGNORECASE
                 )
+
                 if in_items:
                     input_list.append(line.split())
-
                 if out_items:
                     output_list.append(line.split())
 
-        print("Inout List :", input_list)
-        print("Output list", output_list)
+        print("Input List :", input_list)
+        print("Output list :", output_list)
 
         self.input_port = []
         self.output_port = []
 
         # creating list of input and output port with its weight
         for input in input_list:
-            self.input_port.append(input[0]+":"+input[2])
+            self.input_port.append(input[0] + ":" + input[2])
         for output in output_list:
-            self.output_port.append(output[0]+":"+output[2])
+            self.output_port.append(output[0] + ":" + output[2])
 
         print("Output Port List : ", self.output_port)
         print("Input Port List : ", self.input_port)
@@ -145,8 +162,8 @@ class ModelGeneration:
         cfunc = open('cfunc.mod', 'w')
         print("Building content for cfunc.mod file")
 
-        comment = '''/* This is cfunc.mod file auto generated by gen_con_info.py
-        Developed by Fahim, Rahul at IIT Bombay */\n
+        comment = '''/* This is cfunc.mod file auto generated by \
+        model_generation.py \nDeveloped by Fahim, Rahul at IIT Bombay */\n
                 '''
 
         header = '''
@@ -154,14 +171,24 @@ class ModelGeneration:
         #include <math.h>
         #include <string.h>
         #include <time.h>
-        #include <sys/socket.h>
         #include <sys/types.h>
-        #include <netinet/in.h>
-        #include <netdb.h>
         #include <stdlib.h>
         #include <unistd.h>
         #include <errno.h>
+
         '''
+
+        if os.name == 'nt':
+            header += '''
+            #undef BOOLEAN
+            #include<winsock2.h>
+            '''
+        else:
+            header += '''
+            #include <sys/socket.h>
+            #include <netinet/in.h>
+            #include <netdb.h>
+            '''
 
         function_open = (
             '''void cm_''' + self.fname.split('.')[0] + '''(ARGS) \n{''')
@@ -177,7 +204,7 @@ class ModelGeneration:
             // Declaring components of Client
             FILE *log_client = NULL;
             log_client=fopen("client.log","a");
-            int socket_fd, bytes_recieved;
+            int bytes_recieved;
             char send_data[1024];
             char recv_data[1024];
             char *key_iter;
@@ -185,6 +212,11 @@ class ModelGeneration:
             struct sockaddr_in server_addr;
             int sock_port = 5000+PARAM(instance_id);
         '''
+
+        if os.name != 'nt':
+            var_section += '''
+                int socket_fd;
+            '''
 
         temp_input_var = []
         for item in self.input_port:
@@ -261,30 +293,44 @@ class ModelGeneration:
                 int ip_count = 0;
                 char* my_ip = malloc(16);
 
-                char ip_filename[40];
-                sprintf(ip_filename, "/tmp/NGHDL_COMMON_IP_%d.txt", getpid());
+                char ip_filename[100];
+        '''
 
+        if os.name == 'nt':
+            client_setup_ip += '''
+                    sprintf(ip_filename, "''' + \
+                os.getenv('LOCALAPPDATA').replace('\\', '/') + \
+                '''/Temp/NGHDL_COMMON_IP_%d.txt", getpid());
+            '''
+        else:
+            client_setup_ip += '''
+                    sprintf(ip_filename, "/tmp/NGHDL_COMMON_IP_%d.txt",''' \
+                    ''' getpid());
+            '''
+
+        client_setup_ip += '''
                 fptr = fopen(ip_filename, "r");
                 if (fptr)
                 {
-                    char line[20];
-                    while(fscanf(fptr, "%s", line) == 1) {
+                    char line_ip[20];
+                    int line_port;
+                    while(fscanf(fptr, "%s %d", line_ip, &line_port) == 2) {
                         ip_count++;
                     }
 
                     fclose(fptr);
                 }
 
-                if (ip_count < 255) {
+                if (ip_count < 254) {
                     sprintf(my_ip, "127.0.0.%d", ip_count+1);
                 } else {
-                    sprintf(my_ip, "127.0.%d.1", (ip_count+1)%256);
+                    sprintf(my_ip, "127.0.%d.1", (ip_count+3)%256);
                 }
 
                 fptr = fopen(ip_filename, "a");
                 if (fptr)
                 {
-                    fprintf(fptr, "%s\\n", my_ip);
+                    fprintf(fptr, "%s %d\\n", my_ip, sock_port);
                     fclose(fptr);
                 } else {
                     perror("Client - cannot open Common_IP file ");
@@ -296,7 +342,16 @@ class ModelGeneration:
 
         client_fetch_ip = '''
             /* Client Fetch IP Addr */
+        '''
 
+        if os.name == 'nt':
+            client_fetch_ip += '''
+                WSADATA WSAData;
+                SOCKET socket_fd;
+                WSAStartup(MAKEWORD(2, 2), &WSAData);
+            '''
+
+        client_fetch_ip += '''
             char* my_ip = STATIC_VAR(my_ip);
 
             host = gethostbyname(my_ip);
@@ -396,7 +451,18 @@ class ModelGeneration:
             if ( send(socket_fd,send_data,sizeof(send_data),0)==-1)
             {
                 fprintf(stderr, "Client-Failure Sending Message \\n");
-                close(socket_fd);
+        '''
+
+        if os.name == 'nt':
+            send_data += '''
+                    closesocket(socket_fd);
+            '''
+        else:
+            send_data += '''
+                    close(socket_fd);
+            '''
+
+        send_data += '''
                 exit(1);
             }
             else
@@ -429,7 +495,7 @@ class ModelGeneration:
 
         for item in self.output_port:
             sch_output_event.append(
-                 "\t/* Scheduling event and processing them */\n\
+                "\t/* Scheduling event and processing them */\n\
         \tif((key_iter=strstr(recv_data, " + '"' + item.split(':')[0] + ':"'")) != NULL)\n\
         \t{\n\
         \t\twhile(*key_iter++ != ':');\n\
@@ -441,8 +507,8 @@ class ModelGeneration:
         \t\t\telse if(*key_iter=='1')\n\t\t\t{\n\
         \t\t\t\t_op_" + item.split(':')[0] + "[Ii]=ONE;\n\
         \t\t\t}\n\t\t\telse\n\t\t\t{\n\
-        \t\t\t\tfprintf(log_client,\"Unknow value return from server \\n\");\n\
-        \t\t\t\tprintf(\"Client-Unknown value return \\n\");\n\t\t\t}\n\n\
+        \t\t\t\tfprintf(log_client,\"Unknown value return from server \\n\");\
+        \n\t\t\t\tprintf(\"Client-Unknown value return \\n\");\n\t\t\t}\n\n\
         \t\t\tif(ANALYSIS == DC)\n\t\t\t{\n\
         \t\t\t\tOUTPUT_STATE(" + item.split(':')[0] + "[Ii]) = _op_" + item.split(':')[0] + "[Ii];\n\
         \t\t\t}\n\t\t\telse if(_op_" + item.split(':')[0] + "[Ii] != _op_" + item.split(':')[0] + "_old[Ii])\n\
@@ -496,12 +562,31 @@ class ModelGeneration:
         cfunc.write(client_setup_ip)
         cfunc.write("\n")
         cfunc.write("\t\tchar command[1024];\n")
-        cfunc.write(
-            '\t\tsnprintf(command,1024,"' + self.home +
-            '/ngspice-nghdl/src/xspice/icm/ghdl/' +
-            self.fname.split('.')[0] +
-            '/DUTghdl/start_server.sh %d %s &",sock_port,my_ip);'
-        )
+
+        if os.name == 'nt':
+            self.digital_home = self.parser.get('NGHDL', 'DIGITAL_MODEL')
+            self.digital_home = os.path.join(self.digital_home, "ghdl")
+
+            cmd_str2 = "/start_server.sh %d %s & read" + "\\" + "\"" + "\""
+            cmd_str1 = os.path.normpath(
+                "\"" + self.digital_home + "/" +
+                self.fname.split('.')[0] + "/DUTghdl/"
+            )
+            cmd_str1 = cmd_str1.replace("\\", "/")
+
+            cfunc.write(
+                '\t\tsnprintf(command,1024, "start mintty.exe -t ' +
+                '\\"VHDL-Testbench Logs\\" -h always bash.exe -c ' +
+                '\\' + cmd_str1 + cmd_str2 + ', sock_port, my_ip);'
+            )
+        else:
+            cfunc.write(
+                '\t\tsnprintf(command,1024,"' + self.home +
+                '/nghdl-simulator/src/xspice/icm/ghdl/' +
+                self.fname.split('.')[0] +
+                '/DUTghdl/start_server.sh %d %s &", sock_port, my_ip);'
+            )
+
         cfunc.write('\n\t\tsystem(command);')
         cfunc.write("\n\t}")
         cfunc.write("\n")
@@ -533,7 +618,10 @@ class ModelGeneration:
             cfunc.write(item)
 
         # Close socket fd
-        cfunc.write("\tclose(socket_fd);\n\n")
+        if os.name == 'nt':
+            cfunc.write("\tclosesocket(socket_fd);\n\n")
+        else:
+            cfunc.write("\tclose(socket_fd);\n\n")
 
         # close log_client file
         cfunc.write("\tfclose(log_client);")
@@ -633,6 +721,15 @@ class ModelGeneration:
         Vector:              no                          no
         Vector_Bounds:       -                           -
         Null_Allowed:       yes                         yes
+
+        PARAMETER_TABLE:
+        Parameter_Name:     hex_path
+        Description:        "Path of your hex file"                
+        Data_Type:          string                        
+        Default_Value:      "/home/"                                        
+        Vector:              no                          
+        Vector_Bounds:       -                           
+        Null_Allowed:       yes
 
         '''
 
@@ -897,7 +994,7 @@ class ModelGeneration:
             else:
                 process.append(
                     '\t\t\t' + item.split(':')[0] +
-                    ' <= To_Std_Logic('+item.split(':')[0]+'_v'+');\n'
+                    ' <= To_Std_Logic(' + item.split(':')[0] + '_v' + ');\n'
                 )
 
             port_vector_count += 1
@@ -911,20 +1008,20 @@ class ModelGeneration:
                     '\t\t\t' + item.split(':')[0] +
                     '_v := Pack_String_To_Vhpi_String' +
                     '(Convert_SLV_To_String(' +
-                    item.split(':')[0]+'));\n'
+                    item.split(':')[0] + '));\n'
                 )
             else:
                 process.append(
                     '\t\t\t' + item.split(':')[0] +
                     '_v := Pack_String_To_Vhpi_String(To_String(' +
-                    item.split(':')[0]+'));\n'
+                    item.split(':')[0] + '));\n'
                 )
 
             port_vector_count += 1
 
             process.append(
                 '\t\t\tobj_ref := Pack_String_To_Vhpi_String("' +
-                item.split(':')[0]+'");\n'
+                item.split(':')[0] + '");\n'
             )
             process.append(
                 '\t\t\tVhpi_Set_Port_Value(obj_ref,' +
@@ -960,7 +1057,7 @@ class ModelGeneration:
         for item in map:
             testbench.write(item)
 
-        testbench.write("\n\t"+tb_clk)
+        testbench.write("\n\t" + tb_clk)
 
         for item in process_Vhpi:
             testbench.write(item)
@@ -973,30 +1070,53 @@ class ModelGeneration:
     def createServerScript(self):
 
         # ####### Creating and writing components in start_server.sh ####### #
+        self.digital_home = self.parser.get('NGHDL', 'DIGITAL_MODEL')
+        self.digital_home = os.path.join(self.digital_home, "ghdl")
 
         start_server = open('start_server.sh', 'w')
 
         start_server.write("#!/bin/bash\n\n")
         start_server.write(
-            "###This server run ghdl testebench for infinite time till " +
-            "ngspice send END signal to stop it\n\n"
+            "###This server run ghdl testbench for infinite time till " +
+            "Ngspice sends kill signal to stop it\n\n"
         )
-        start_server.write(
-            "cd "+self.home+"/ngspice-nghdl/src/xspice/icm/ghdl/" +
-            self.fname.split('.')[0]+"/DUTghdl/\n"
-        )
+
+        if os.name == 'nt':
+            pathstr = self.digital_home + "/" + \
+                self.fname.split('.')[0] + "/DUTghdl/"
+            pathstr = pathstr.replace("\\", "/")
+            start_server.write("cd " + pathstr + "\n")
+        else:
+            start_server.write("cd " + self.digital_home +
+                               "/" + self.fname.split('.')[0] + "/DUTghdl/\n")
+
         start_server.write("chmod 775 sock_pkg_create.sh &&\n")
         start_server.write("./sock_pkg_create.sh $1 $2 &&\n")
         start_server.write("ghdl -i *.vhdl &&\n")
         start_server.write("ghdl -a *.vhdl &&\n")
-        start_server.write("ghdl -a "+self.fname+" &&\n")
+        start_server.write("ghdl -a " + self.fname + " &&\n")
         start_server.write(
-            "ghdl -a "+self.fname.split('.')[0]+"_tb.vhdl  &&\n"
+            "ghdl -a " + self.fname.split('.')[0] + "_tb.vhdl  &&\n"
         )
-        start_server.write(
-            "ghdl -e -Wl,ghdlserver.o " + self.fname.split('.')[0] + "_tb &&\n"
-        )
-        start_server.write("./"+self.fname.split('.')[0]+"_tb")
+
+        if os.name == 'nt':
+            start_server.write("ghdl -e -Wl,ghdlserver.o " +
+                               "-Wl,libws2_32.a " +
+                               self.fname.split('.')[0] + "_tb &&\n")
+            start_server.write("./" + self.fname.split('.')[0] + "_tb.exe")
+        else:
+            start_server.write("ghdl -e -Wl,ghdlserver.o " +
+                               self.fname.split('.')[0] + "_tb &&\n")
+            start_server.write(
+                "./" +
+                self.fname.split('.')[0] +
+                "_tb --vcd=" +
+                self.fname.split('.')[0] +
+                "_tb.vcd\n")
+            start_server.write(
+                "gtkwave " +
+                self.fname.split('.')[0] +
+                "_tb.vcd 2>/dev/null")
 
         start_server.close()
